requirements.txt
customtkinter
requests
python-dotenv
moviepy
google-genai
